+++ projects/core/server/expressBridge.ts
@@ -1087,7 +1087,7 @@ export class EntityQueueStorage implements QueueStorage {
   }
 
   async getJobInfo(queuedJobId: string): Promise<queuedJobInfo> {
-    let q = await this.repo.findId(queuedJobId, { useCache: false })
+    let q = await this.repo.findId(queuedJobId)
     let lastProgress: Date = undefined
     return {
       userId: q.userId,